Create new hierarchy for spatial functions
