+++ python/cuspatial/cuspatial/core/spatial/filter.py
